Move route handler tests to tests/routes/
Tests need the database and helpers, so by the convention in
tests/README.md they belong under tests/, not colocated.
